backend/pyproject.toml
[project]
name = "construction-suite-backend"
version = "0.1.0"
description = "Construction management backend (Django + DRF)"
requires-python = ">=3.11"
dependencies = [
  "Django>=5.0,<6.0",
  "djangorestframework>=3.15,<4.0",
  "djangorestframework-simplejwt>=5.3,<6.0",
  "drf-spectacular>=0.27,<1.0",
  "django-filter>=24.2,<25.0",
  "django-cors-headers>=4.3,<5.0",
  "psycopg[binary]>=3.1,<4.0",
  "gunicorn>=22.0,<23.0",
]

[tool.black]
line-length = 100
target-version = ["py311"]

[tool.isort]
profile = "black"
line_length = 100
